requirements.txt
google-adk==1.0.0
fastapi==0.115.0
uvicorn==0.34.0
cryptography==44.0.2
itsdangerous==2.2.0
google-cloud-bigquery==3.25.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
mcp==1.5.0
